@@ -1,6 +1,14 @@
-import React from "react";
+import React, { Component } from "react";
 
-const Comments = React.createClass({
+class Comments extends Component {
+  constructor(props) {
+    super(props);
+    this.renderComment = this.renderComment.bind(this);
+    this.handleSubmit = this.handleSubmit.bind(this);
+    this.commentForm = React.createRef();
+    this.author = React.createRef();
+    this.comment = React.createRef();
+  }
   renderComment(comment, i) {
     return (
       <div className="comment" key={i}>
@@ -20,31 +28,31 @@ const Comments = React.createClass({
         </p>
       </div>
     );
-  },
+  }
   handleSubmit(e) {
     e.preventDefault();
     const { postId } = this.props.params;
-    const author = this.refs.author.value;
-    const comment = this.refs.comment.value;
+    const author = this.author.current.value;
+    const comment = this.comment.current.value;
     this.props.addComment(postId, author, comment);
-    this.refs.commentForm.reset();
-  },
+    this.commentForm.current.reset();
+  }
   render() {
     return (
       <div className="comments">
         {this.props.postComments.map(this.renderComment)}
         <form
-          ref="commentForm"
+          ref={this.commentForm}
           className="comment-form"
           onSubmit={this.handleSubmit}
         >
-          <input type="text" ref="author" placeholder="author" />
-          <input type="text" ref="comment" placeholder="comment" />
+          <input type="text" ref={this.author} placeholder="author" />
+          <input type="text" ref={this.comment} placeholder="comment" />
           <input type="submit" hidden />
         </form>
       </div>
     );
-  },
-});
+  }
+}
 
 export default Comments;
