@@ -2,16 +2,14 @@ import React from "react";
 
 import { Link } from "react-router";
 
-const Main = React.createClass({
-  render() {
-    return (
-      <div>
-        <h1>
-          <Link to="/">daGram</Link>
-        </h1>
-        {React.cloneElement(this.props.children, this.props)}
-      </div>
-    );
-  },
-});
+function Main(props) {
+  return (
+    <div>
+      <h1>
+        <Link to="/">daGram</Link>
+      </h1>
+      {React.cloneElement(props.children, props)}
+    </div>
+  );
+}
 export default Main;
