@@ -1,15 +1,13 @@
 import React from "react";
 
 import Photo from "./Photo";
-const PhotoGrid = React.createClass({
-  render() {
-    return (
-      <div className="photo-grid">
-        {this.props.posts.map((post, i) => (
-          <Photo key={i} {...this.props} post={post} i={i} />
-        ))}
-      </div>
-    );
-  },
-});
+function PhotoGrid(props) {
+  return (
+    <div className="photo-grid">
+      {props.posts.map((post, i) => (
+        <Photo key={i} {...props} post={post} i={i} />
+      ))}
+    </div>
+  );
+}
 export default PhotoGrid;
