@@ -2,22 +2,20 @@ import React from "react";
 import Photo from "./Photo";
 import Comments from "./Comments";
 
-const Single = React.createClass({
-  render() {
-    const { postId } = this.props.params;
-    // index of the post
-    const i = this.props.posts.findIndex((post) => post.code === postId);
-    // get the post
-    const post = this.props.posts[i];
-    // get the comments
-    const postComments = this.props.comments[postId] || [];
+function Single(props) {
+  const { postId } = props.params;
+  // index of the post
+  const i = props.posts.findIndex((post) => post.code === postId);
+  // get the post
+  const post = props.posts[i];
+  // get the comments
+  const postComments = props.comments[postId] || [];
 
-    return (
-      <div className="single-photo">
-        <Photo i={i} post={post} {...this.props} />
-        <Comments postComments={postComments} {...this.props} />
-      </div>
-    );
-  },
-});
+  return (
+    <div className="single-photo">
+      <Photo i={i} post={post} {...props} />
+      <Comments postComments={postComments} {...props} />
+    </div>
+  );
+}
 export default Single;
