@@ -1,4 +1,4 @@
-import { createStore, compose } from "redux";
+import { createStore } from "redux";
 import { syncHistoryWithStore } from "react-router-redux";
 import { browserHistory } from "react-router";
 
